requirements.txt
bigmodelvis==0.0.1
deepspeed==0.9.3
easydict==1.13
einops==0.6.1
fastapi==0.109.2
gradio==3.35.2
gradio_client==0.2.9
ipdb==0.13.13
numpy==1.24.3
peft==0.3.0
Pillow==9.5.0
PyYAML==6.0
requests==2.31.0
sentence-transformers==2.6.1
sentencepiece==0.1.99
setuptools==65.5.1
tensorboard==2.14.0
timm==0.6.7
torch==1.13.1
torchvision==0.14.1
tqdm==4.64.1
transformers==4.35.0
uvicorn==0.27.1
